[step-6]: 세션 기능 수정 요청별로 세션 인스턴스를 생성할 수 있게 하여, 여러 정보를 저장할 수 있게 함
